Actualizando Paciente

+++ clinicaOdontologica/src/main/java/com/backend/clinicaOdontologica/dao/impl/PacienteDaoH2.java
@@ -35,14 +35,14 @@ public class PacienteDaoH2 implements IDao<Paciente> {
             preparedStatement.setString(2, paciente.getApellido());
             preparedStatement.setInt(3, paciente.getDni());
             preparedStatement.setDate(4, Date.valueOf(paciente.getFechaIngreso()));
-            preparedStatement.setInt(5, domicilioRegistrado.getId());
+            preparedStatement.setLong(5, domicilioRegistrado.getId());
             preparedStatement.execute();
 
             pacienteRegistrado = new Paciente(paciente.getNombre(), paciente.getApellido(), paciente.getDni(), paciente.getFechaIngreso(), domicilioRegistrado);
 
             ResultSet resultSet = preparedStatement.getGeneratedKeys();
             while (resultSet.next()) {
-                pacienteRegistrado.setId(resultSet.getInt("id"));
+                pacienteRegistrado.setId(resultSet.getLong("id"));
             }
 
             connection.commit();
@@ -151,6 +151,6 @@ public class PacienteDaoH2 implements IDao<Paciente> {
 
         Domicilio domicilio = new DomicilioDaoH2().buscarPorId(resultSet.getInt("domicilio_id"));
 
-        return new Paciente(resultSet.getInt("id"), resultSet.getString("nombre"), resultSet.getString("apellido"), resultSet.getInt("dni"), resultSet.getDate("fecha").toLocalDate(), domicilio);
+        return new Paciente(resultSet.getLong("id"), resultSet.getString("nombre"), resultSet.getString("apellido"), resultSet.getInt("dni"), resultSet.getDate("fecha").toLocalDate(), domicilio);
     }
 }
