+++ clinicaOdontologica/src/main/java/com/backend/clinicaOdontologica/service/IPacienteService.java
@@ -9,5 +9,5 @@ public interface IPacienteService {
 
     List<PacienteSalidaDto> listarPacientes();
 
-    PacienteSalidaDto buscarPacientePorId(int id);
+    PacienteSalidaDto buscarPacientePorId(Long id);
 }
